SY-4099: Mirror setup/verify and setup version by tag and minimize setup_versions
