+++ integration/tests/migration_tests.json
@@ -4,13 +4,13 @@
     {
       "sequence_name": "all",
       "tests": [
-        { "case": "migration/channels" },
-        { "case": "migration/channels_calc" },
-        { "case": "migration/ranges" },
-        { "case": "migration/rbac" },
-        { "case": "migration/task_modbus" },
-        { "case": "migration/task_ni" },
-        { "case": "migration/task_opc" }
+        { "case": "migration/channels_verify" },
+        { "case": "migration/channels_calc_verify" },
+        { "case": "migration/ranges_verify" },
+        { "case": "migration/rbac_verify" },
+        { "case": "migration/task_modbus_verify" },
+        { "case": "migration/task_ni_verify" },
+        { "case": "migration/task_opc_verify" }
       ]
     }
   ]
